requirements.txt
numpy==1.21.2
matplotlib==3.4.3
tqdm==4.62.3
pandas==1.3.4
scipy==1.7.1
pytorch==1.10.0